==> 01_data_preparation/01_pull_variables.Rmd <==
---
title: 01_pull_variables
---

```{r set_libraries, include=FALSE, tidy=TRUE}

require(tidyverse)

#Set path for raw data files
path <- '/Users/ab3377/Library/CloudStorage/GoogleDrive-ab3377@yale.edu/My Drive/CANDLab/Projects/ELAdata'
```


```{r}
#Read in family variable
rel <- read_tsv(paste(path, 'acspsw03.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,rel_family_id)
```

```{r}
#Read in family substance use summary scores
fhx <- read_tsv(paste(path, 'abcd_fhxssp01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,famhx_ss_fath_prob_alc_p,famhx_ss_moth_prob_alc_p,famhx_ss_fath_prob_dg_p,
                famhx_ss_moth_prob_dg_p)
```

```{r}
##Read in parent demographics

pdemo <- read_tsv(paste(path,'pdem02.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,demo_prim,demo_prnt_marital_v2,demo_prnt_ed_v2,demo_prtnr_ed_v2,demo_comb_income_v2,
                demo_fam_exp1_v2,demo_fam_exp2_v2,demo_fam_exp3_v2,demo_fam_exp4_v2,demo_fam_exp5_v2,
                demo_fam_exp6_v2,demo_fam_exp7_v2)
```

```{r}
##Read in CRPBI
crpbi <- read_tsv(paste(path, 'crpbi01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,crpbi_parent1_y,crpbi_caregiver12_y,crpbi_parent2_y,crpbi_caregiver13_y,
                crpbi_parent3_y, crpbi_caregiver14_y,crpbi_parent4_y,crpbi_caregiver15_y,crpbi_parent5_y,
                crpbi_caregiver16_y)
```

```{r}
##Read in parent report family environment scale
fes02 <- read_tsv(paste(path, 'fes02.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,fam_enviro1_p,fam_enviro2r_p,fam_enviro3_p,fam_enviro4r_p,fam_enviro5_p,
                fam_enviro6_p,fam_enviro7r_p,fam_enviro8_p,fam_enviro9r_p)

#Note: NAs for many items (below), unclear why. Omitted these variables.
#Fes_31_p,fes_32_p,fes_51_p,fes_52r_p,fes_1_p,fes_72r_p,fes_2r_p,fes_81_p,fes_19_p,fes_29r_p,fes_39_p,fes_42_p,fes_49r_p,fes_59_p,fes_69_p, fes_79r_p,fes_9_p
```

```{r}
##Read in youth report family environment scale
fes01 <- read_tsv(paste(path, 'abcd_fes01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,fes_youth_q1,fes_youth_q2,fes_youth_q3,fes_youth_q4,fes_youth_q5,fes_youth_q6,
                fes_youth_q7,fes_youth_q8,fes_youth_q9)
```

```{r}
##Read in ksads trauma, parent interview
ptsd <- read_tsv(paste(path, 'abcd_ptsd01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,ksads_ptsd_raw_754_p,ksads_ptsd_raw_755_p,ksads_ptsd_raw_756_p,ksads_ptsd_raw_757_p,
                ksads_ptsd_raw_758_p,ksads_ptsd_raw_759_p,ksads_ptsd_raw_760_p,ksads_ptsd_raw_761_p,
                ksads_ptsd_raw_762_p,ksads_ptsd_raw_763_p,ksads_ptsd_raw_764_p,ksads_ptsd_raw_765_p,
                ksads_ptsd_raw_766_p,ksads_ptsd_raw_767_p,ksads_ptsd_raw_768_p,ksads_ptsd_raw_769_p,
                ksads_ptsd_raw_770_p)
```

```{r}
##Read in parental monitoring
pmq <- read_tsv(paste(path, 'pmq01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,parent_monitor_q1_y,parent_monitor_q2_y,parent_monitor_q3_y,parent_monitor_q4_y,
                parent_monitor_q5_y)
```

```{r}
#Read in neighborhood safety and crime, parents
pnscss <- read_tsv(paste(path, 'abcd_sscep01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,nsc_p_ss_mean_3_items) 
```

```{r}
##Read in neighborhood safety and crime, youth
ynsc <- read_tsv(paste(path, 'abcd_nsc01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,neighborhood_crime_y)
```

```{r}
##Read in ASR (parent psychopathology)
asr <- read_tsv(paste(path, 'abcd_asrs01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,asr_scr_anxdisord_r,asr_scr_somaticpr_r,asr_scr_depress_r,asr_scr_avoidant_r,
                asr_scr_adhd_r,asr_scr_antisocial_r,asr_scr_inattention_r,asr_scr_hyperactive_r)
```

```{r}
#Read in ADI data file
ADI <- read_tsv(paste(path, 'abcd_rhds01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,reshist_addr1_adi_wsum)

```

```{r merge all data frames}
ELAdata <- full_join(fhx, asr)%>%
  full_join(pdemo) %>%
  full_join(crpbi) %>%
  full_join(fes01)%>%
  full_join(fes02)%>%
  full_join(ptsd)%>%
  full_join(pmq)%>%
  full_join(pnscss)%>%
  full_join(ynsc)%>%
  full_join(rel)%>%
  full_join(ADI)
```

```{r}
write.csv(ELAdata, file = "ELAdata.csv", row.names = FALSE)
```

```{r select one relative from each family}

#read in tsv file with IDs (note: separately, in python, we selected one relative from each family, resulting in n = 9854)

ELA_norel_ID <- read_tsv('ELAdata_norel_ID.tsv', col_names=TRUE)

#merge with ELA data
ELA_data_norel <- merge(ELAdata, ELA_norel_ID[, c('subjectkey')], by = "subjectkey")

```


```{r}
##Read in and merge demographics
demo <- read_tsv(paste(path, 'abcd_lt01.txt', sep='/'), col_names=TRUE) %>%
  filter(eventname == "baseline_year_1_arm_1") %>% 
  dplyr::select(subjectkey,interview_age, sex)

ELA_data_norel <- merge(ELA_data_norel, demo, by = "subjectkey")
```

```{r remove family ID, no longer needed}

ELA_data_norel <- ELA_data_norel %>%
    dplyr::select(-rel_family_id)
```

```{r convert variables to numeric and recode missing values}

ELA_data_norel <- ELA_data_norel %>%
  mutate_at(vars(2:79), as.numeric)

#Replace 777 and 999 as na

ELA_data_norel <- ELA_data_norel %>%
  mutate_at(vars(2:79), ~replace(., . == 777 | . == 999, NA))

#confirm that there are no 777 999 values
has_777_or_999 <- any(ELA_data_norel == 777 | df_updated == 999, na.rm = TRUE)
if (has_777_or_999) {
  print("The dataframe contains values of 777 or 999.")
} else {
  print("The dataframe does not contain values of 777 or 999.")
}
```

```{r check degree of missing data across variables}

ELA_var_sums_NA <- as.data.frame(sapply(ELA_data_norel, function(x) sum(is.na(x))))
ELA_var_sums_NA
```

```{r recode primary caregiver and marital status variables}
#recode demo prim, 1=bio separation from parent, 0=lives with bio parents

ELA_data_norel$demo_prim_biosep <- ifelse(ELA_data_norel$demo_prim >= 3,1,0)

#recode parent marital status, coding 1 and 6 are now = 0 (no parental separation) otherwise 1 (parental separation)

ELA_data_norel$demo_parent_sep <- ifelse(ELA_data_norel$demo_prnt_marital_v2 == 1 |    ELA_data_norel$demo_prnt_marital_v2 == 6, 0, 1)

#remove original variables
ELAdata_forFA <- ELA_data_norel %>%
   dplyr::select(-demo_prim, -demo_prnt_marital_v2)

```

```{r save out data}
write.csv(ELAdata_forFA, file = "ELAdata_forFA.csv", row.names = FALSE)
```
